requirements.txt
django==1.11.1
scrapy==1.5.1
djangorestframework==3.6.3